@@ -4,9 +4,9 @@
 // Original: https://github.com/ahmadawais/shades-of-purple-vscode/
 // Converted automatically using ./tools/themeFromVsCode and then customized manually.
 
-/*:: import type { PrismTheme } from '../src/types' */
+import type { PrismTheme } from "../types";
 
-var theme /*: PrismTheme */ = {
+var theme: PrismTheme = {
   plain: {
     color: "#9EFEFF",
     backgroundColor: "#2D2A55"
@@ -96,4 +96,4 @@ var theme /*: PrismTheme */ = {
   ]
 };
 
-module.exports = theme;
+export default theme;
