@@ -1,129 +1,96 @@
 // @flow
 // Converted automatically using ./tools/themeFromVsCode
 
-/*:: import type { PrismTheme } from "../src/types" */
+import type { PrismTheme } from "../types";
 
-var theme /*: PrismTheme */ = {
+var theme: PrismTheme = {
   plain: {
     color: "#bfc7d5",
     backgroundColor: "#292d3e"
   },
   styles: [
     {
-      types: [
-        "comment"
-      ],
+      types: ["comment"],
       style: {
         color: "rgb(105, 112, 152)",
         fontStyle: "italic"
       }
     },
     {
-      types: [
-        "string"
-      ],
+      types: ["string"],
       style: {
         color: "rgb(195, 232, 141)"
       }
     },
     {
-      types: [
-        "number"
-      ],
+      types: ["number"],
       style: {
         color: "rgb(247, 140, 108)"
       }
     },
     {
-      types: [
-        "builtin",
-        "char",
-        "constant",
-        "function"
-      ],
+      types: ["builtin", "char", "constant", "function"],
       style: {
         color: "rgb(130, 170, 255)"
       }
     },
     {
-      types: [
-        "punctuation",
-        "selector"
-      ],
+      types: ["punctuation", "selector"],
       style: {
         color: "rgb(199, 146, 234)"
       }
     },
     {
-      types: [
-        "variable"
-      ],
+      types: ["variable"],
       style: {
         color: "rgb(191, 199, 213)"
       }
     },
     {
-      types: [
-        "class-name",
-        "attr-name"
-      ],
+      types: ["class-name", "attr-name"],
       style: {
         color: "rgb(255, 203, 107)"
       }
     },
     {
-      types: [
-        "tag"
-      ],
+      types: ["tag"],
       style: {
         color: "rgb(255, 85, 114)"
       }
     },
     {
-      types: [
-        "operator"
-      ],
+      types: ["operator"],
       style: {
         color: "rgb(137, 221, 255)"
       }
     },
     {
-      types: [
-        "boolean"
-      ],
+      types: ["boolean"],
       style: {
         color: "rgb(255, 88, 116)"
       }
     },
     {
-      types: [
-        "keyword"
-      ],
+      types: ["keyword"],
       style: {
         fontStyle: "italic"
       }
     },
     {
-      types: [
-        "doctype"
-      ],
+      types: ["doctype"],
       style: {
         color: "rgb(199, 146, 234)",
         fontStyle: "italic"
       }
     },
     {
-      types: [
-        "namespace"
-      ],
+      types: ["namespace"],
       style: {
         color: "rgb(178, 204, 214)"
       }
     },
     {
-      types: [
-        "url"
-      ],
+      types: ["url"],
       style: {
         color: "rgb(221, 221, 221)"
       }
